requirements.txt
asgiref==3.8.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.17.0
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
psycopg2==2.9.10
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.1
uvicorn==0.34.0
virtualenv==20.29.1
whitenoise==6.9.0
